Added DateTime properties

+++ src/xfnet/XfModels/Date.cs
@@ -5,12 +5,12 @@ namespace xfnet.XfModels
     public class XfDate
     {
         [JsonProperty("year")]
-        public long? Year { get; set; }
+        public int? Year { get; set; }
 
         [JsonProperty("month")]
-        public long? Month { get; set; }
+        public int? Month { get; set; }
 
         [JsonProperty("day")]
-        public long? Day { get; set; }
+        public int? Day { get; set; }
     }
 }
